Refactor test packages names
