+++ auction-app-backend/src/test/java/com/atlantbh/auctionappbackend/controllerTest/AuthControllerNonMockTest.java
@@ -1,4 +1,4 @@
-package com.atlantbh.auctionappbackend.authControllerTest;
+package com.atlantbh.auctionappbackend.controllerTest;
 
 import com.atlantbh.auctionappbackend.domain.User;
 import com.atlantbh.auctionappbackend.repository.UserRepository;
