+++ auction-app-backend/src/test/java/com/atlantbh/auctionappbackend/repostitoryTest/UserRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.atlantbh.auctionappbackend.userRepostitoryTest;
+package com.atlantbh.auctionappbackend.repostitoryTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -10,10 +10,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.test.annotation.Rollback;
 
-import java.util.Arrays;
 import java.util.List;
 
 @DataJpaTest
